refactor

+++ app/src/main/java/com/example/weatherapp/utils/TimeFormatter.kt
@@ -6,19 +6,19 @@ object TimeFormatter {
 
     fun getDateFromUnixTime(unixTime: Long): String{
         val sdf = java.text.SimpleDateFormat("dd.MM.yyyy HH:mm", Locale.getDefault())
-        val date = Date(unixTime.toLong() * 1000)
+        val date = Date(unixTime * 1000)
         return sdf.format(date)
     }
 
     fun getTimeFromUnixTime(unixTime: Long): String{
         val sdf = java.text.SimpleDateFormat("HH:mm", Locale.getDefault())
-        val date = Date(unixTime.toLong()*1000)
+        val date = Date(unixTime *1000)
         return sdf.format(date)
     }
 
     fun getDMFromUnixTime(unixTime: Long): String{
         val sdf = java.text.SimpleDateFormat("dd.MM", Locale.getDefault())
-        val date = Date(unixTime.toLong() * 1000)
+        val date = Date(unixTime * 1000)
         return sdf.format(date)
     }
